Modified   examples/BluetoothPlayer.h Modified   examples/Passthrough.h Modified   examples/SineWavePlayer.h Modified   include/Wave/BiquadFilter.h Modified   include/Wave/HighPassFilter.h Modified   include/Wave/LowPassFilter.h Modified   include/Wave/NoiseGate.h Modified   include/Wave/Oscillator.h Modified   include/Wave/ToneControl.h Modified   include/Wave/Wave.h Modified   include/Wave/WaveShaper.h Added      include/Task.h

+++ include/Wave/BiquadFilter.h
@@ -40,7 +40,7 @@ public:
 				m_States[i].Coef[j] = Values[(5 * i) + j];
 	}
 
-	float Process(float Value) override
+	double Process(double Value) override
 	{
 		for (uint8 i = 0; i < m_StateCount; ++i)
 		{
