+++ include/Wave/Wave.h
@@ -7,9 +7,9 @@
 class Wave
 {
 public:
-	virtual float Process(float Value) = 0;
+	virtual double Process(double Value) = 0;
 
-	virtual void ProcessBuffer(float *Buffer, uint16 Count)
+	virtual void ProcessBuffer(double *Buffer, uint16 Count)
 	{
 		for (uint16 i = 0; i < Count; ++i)
 			Buffer[i] = Process(Buffer[i]);
